requirements.txt
Babel==2.3.4
Django==1.8.17
Django-Select2==5.3.0
amqp==1.4.7
anyjson==0.3.3
argparse==1.2.1
billiard==3.3.0.21
celery==3.1.19
coverage==4.0.2
django-allauth==0.24.1
django-annoying==0.8.4
django-appconf==1.0.1
django-braces==1.11.0
django-cacheops==2.4.3
django-compressor==1.5
django-crispy-forms==1.5.2
django-datatable-view==0.8.3
django-debug-toolbar==1.4
django-js-reverse==0.6.1
django-jsonview==0.5.0
django-phonenumber-field==1.1.0
django-randomslugfield==0.3.0
django-redis==4.3.0
django-scheduler==0.7.5
django-suit==0.2.15
django-suit-redactor==0.0.2
django-tables2==1.10.0
funcy==1.6
future==0.15.2
icalendar==3.9.1
kombu==3.0.29
msgpack-python==0.4.6
oauthlib==1.0.3
oneapi-python==0.3
pdfkit==0.5.0
phonenumbers==7.1.1
phonenumberslite==8.0.0
psycopg2==2.6.1
python-dateutil==2.5.0
python-openid==2.2.5
pytz==2016.10
redis==2.10.5
requests==2.8.1
requests-oauthlib==0.5.0
six==1.10.0
sqlparse==0.1.18
vobject==0.8.1rc0
wsgiref==0.1.2
